Convert to use a central redux store
This is overkill for what is presently in the repository, but will serve
a more complete application well. The application will have one central
state, and pages can interact with/update that state. It enables sharing
common state, and centralizes API calls/state logic using React with
Redux and some helpers for aynchronous calls.