Rework the development proxy setup with .env files
Use the simpler builtin proxy, and decide upon what prefix to use in the
web application using environment variables. This keeps the current /qs
prefix, but that can be overridden using .env.local or similar. Added
some documentation to explain how to configure this, and verified using
a local queueserver with no prefix override.