Added two more buttons for environments and queues
This is hackish, but I want to get something working and see what we
might be missing. It adds an open/close environment button and a
start/stop queue button. It works in the basic sense, but needs more
state in order to report and work as expected. Still, as a stake in the
ground I think this shows some of the interactions we are after.