Add an editable plan parameter, feed it through
This is the very minimal version of the editable parameter, where a
single number is fed through from the interface to the RESTful call. For
a more complete solution the scope of editable fields must grow, and the
data structure to pass them through should likely be an array of
key-value pairs. It works, and shows the basic flow of data though.