Some very initial sketching out of BMM's main plan
This explores a custom interface for the main BMM plan, not complete but
as discussed today. We agreed on more space for prep, comments, possibly
a spot for the file name prefix, radio buttons for mode, select for edge
and element, and adding the regionds section where four regions would be
the default.