Added in bounds to the form
This is currently hard-coded to four regions, it at least needs to be
geneeralized so that it defaults to four regions but the number can be
changed. It also needs to actually submit somewhere, and we should
consider how we might store the last settings between submissions so
that the user doesn't need to start again for each submission.